refactor: diplomacy wave 5 slice B dossier dialogue/evidence splits (#4130)
Split event_dialogue and evidence_rules into focused modules under the
400-line ratchet and remove both from the lib grandfather allowlist.
