+++ SPEC/program/repo-lint.md
@@ -207,7 +207,7 @@
 | `tool/check_diplomacy_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical Dart LOC (≤2488 after wave-4 densify; ≤2500 band met); rule `repo.diplomacy_test_support_loc` (Refs #4028, #4037) |
 | `tool/check_diplomacy_test_no_src_imports.dart` | Forbid `package:colonizethis_diplomacy/src/` imports under `packages/colonizethis_diplomacy/test/**` except the allowlisted `diplomacy_phase_types_split_test.dart` deep-import gate; tests MUST use the public barrel or `colonizethis_diplomacy_test_support` (Refs #3837); rule `repo.diplomacy_test_no_src_imports` |
 | `tool/check_diplomacy_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_diplomacy/test/**` exceeds 400 physical lines (aligns with economy test file cap); rule `repo.diplomacy_test_file_size` |
-| `tool/check_diplomacy_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist for six lib hotspots during transition; generated suffixes excluded); rule `repo.colonizethis_diplomacy_lib_file_size` (Refs #4130) |
+| `tool/check_diplomacy_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist for four resolver lib hotspots during transition; generated suffixes excluded); rule `repo.colonizethis_diplomacy_lib_file_size` (Refs #4130) |
 | `tool/check_diplomacy_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist for `evidence_rules_scenarios.dart` during densify; generated suffixes excluded); rule `repo.colonizethis_diplomacy_test_support_file_size` (Refs #4130) |
 | `tool/check_diplomacy_test_support_package_only.dart` | Forbid package-local fixture modules under `packages/colonizethis_diplomacy/test/support/**` and new `*_test_support.dart` shims under `packages/colonizethis_diplomacy/test/**`; shared fixtures MUST live in `colonizethis_diplomacy_test_support` (Refs #3837); rule `repo.diplomacy_test_support_package_only` |
 | `tool/check_ai_dedup_weight_scale_clamp.dart` | Forbid the inline weight-scale clamp idiom (`<= 0.0) { return 0; }` guard + `final clamped = X > 1.0 ? 1.0 : X;` + `return (Y * clamped).round();`) under `packages/colonizethis_ai/lib/**` outside the canonical helper file `src/planning/planning_helpers.dart`; filter/scoring code MUST call the shared `scaleWeightedBonus(weight, baseConstant)` helper (Refs #3278); rule `repo.ai_dedup_weight_scale_clamp` |
@@ -381,7 +381,7 @@ Large `app/test/**` widget/screen suites that are split into numbered fragments 
 
 **World test file size (Refs #4125):** `repo.colonizethis_world_test_file_size` fails when any non-support Dart file under `packages/colonizethis_world/test/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist. Files under `test/world_test_support/` are excluded (governed by builder-usage gates).
 
-**Diplomacy lib file size (Refs #4130):** `repo.colonizethis_diplomacy_lib_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (six lib hotspots during wave-5 transition slices).
+**Diplomacy lib file size (Refs #4130):** `repo.colonizethis_diplomacy_lib_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (four resolver lib hotspots during wave-5 transition slices; dossier dialogue/evidence modules split in slice B).
 
 **Diplomacy test support file size (Refs #4130):** `repo.colonizethis_diplomacy_test_support_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy_test_support/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (`evidence_rules_scenarios.dart` during densify slices).
